requirements.txt
boto3==1.26.99
botocore==1.29.99
configparser==5.3.0
PyGithub==1.58.1
debugpy==1.6.4
aiohttp==3.9.2
PyYAML==6.0